Move, optimize and refactor decodeFsst function (#46)
Move function into `index.ts` to remove overhead and optimized /
refactored along the way
